Refactor pixel format code into pixel.go
